Refactor TagsContainsKeywordsPredicate into tag package
