+++ src/main/java/seedu/address/model/tag/TagsContainsKeywordsPredicate.java
@@ -1,8 +1,9 @@
-package seedu.address.model.person;
+package seedu.address.model.tag;
 
 import java.util.List;
 import java.util.function.Predicate;
 import seedu.address.commons.util.StringUtil;
+import seedu.address.model.person.Person;
 
 /**
  * Tests that a {@code Person}'s {@code Tags} matches any of the keywords given.
